+++ app/specific/Main.js
@@ -24,10 +24,10 @@ var Main_isTV;
 var Main_isDebug = false;
 
 var Main_stringVersion = '3.0';
-var Main_stringVersion_Min = '.294';
-var Main_version_java = 293;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
-var Main_minversion = 'January 09 2020';
-var Main_version_web = 562;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
+var Main_stringVersion_Min = '.295';
+var Main_version_java = 295;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
+var Main_minversion = 'January 15 2020';
+var Main_version_web = 563;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
 var Main_versionTag = Main_stringVersion + Main_stringVersion_Min + '-' + Main_minversion;
 
 var Main_cursorYAddFocus = -1;
@@ -633,6 +633,13 @@ function Main_Changelog() {
         STR_DIV_LINK + STR_ABOUT_CHANGELOG + '</div><br><br>';
 
     var changelogObj = [
+        {
+            title: "Apk Version 3.0.295 and Web Version January 15 2020",
+            changes: [
+                "Add OLED Burn in protection option to Settings -> Content customization's, sorting, auto refresh, timeouts and related",
+                "General performance improves and bug fixes"
+            ]
+        },
         {
             title: "Web Version January 09 2020",
             changes: [